Git commit to fix a bug when reading wave data and to add a search function for subchunk id tags.
The bug was an issue with how each byte was read from the file, where a byte with value FF would be set as 00.
This was because the variable that was set by fgetc was signed, where as bytes should be read as an unsigned char.

The search function loops through the wave file and attempts to find the subchunk id tag.
This is better for the data subchunk as there may be extra parameters the code will not read.