Git commit to redesign the layout of Write WAV File library.
The commit has brought about a number of changes, most notably the combining of wave_file.c and wave.c into a single file so the functions available are easier to see.

Reading wave files has been split into multiple methods that will cause the process to terminate if a given chunk is not found.

Comments have been added for every functions to make clear what every component does.
A few of the functions have been renamed for clarity, and the wave and wave_header types have been made lower case and underscore for consistency.

A new field has been added to the wave type - __is_valid. This is general bool value that determines whether it is safe to read and write data to/from the wave.